Set the global locale at start
